Runs without collisions for 200 seconds. Nice job on algorithm, Sean.
1) Moved collision management into process A since it controls movement. Did NOT
move movement to C... too much work. Rather, before any moves occur, run collision
management. Then transmit data to B and C as usual.

2) Created messaging interface to push data across the processes easily. The
arrays were fine for P1, but too much work to figure out what was happening and
were not a uniform interface. Now a Message wraps a set of Positions which is
a collection of Position objects which describes trains.

3) Added ProcessMsgShuttle that can move messages at a specific frequency
between two processes. This will be useful in P3 when messages need to be sent
from process C to control the trains.

4) Changed namespace from com.jeff to com.rtosProject2.